Move RFC tests to common location
