+++ jetty-core/jetty-tests/jetty-test-rfc/src/test/resources/jetty-logging.properties
@@ -1,0 +1,15 @@
+#org.eclipse.jetty.LEVEL=DEBUG
+org.eclipse.jetty.jmx.LEVEL=INFO
+#org.eclipse.jetty.client.LEVEL=DEBUG
+#org.eclipse.jetty.fcgi.LEVEL=DEBUG
+#org.eclipse.jetty.proxy.LEVEL=DEBUG
+#org.eclipse.jetty.http2.LEVEL=DEBUG
+org.eclipse.jetty.http2.hpack.LEVEL=INFO
+#org.eclipse.jetty.http2.client.LEVEL=DEBUG
+#org.eclipse.jetty.http3.LEVEL=DEBUG
+org.eclipse.jetty.http3.qpack.LEVEL=INFO
+#org.eclipse.jetty.quic.LEVEL=DEBUG
+org.eclipse.jetty.quic.quiche.foreign.LEVEL=INFO
+org.eclipse.jetty.quic.quiche.jna.LEVEL=INFO
+#org.eclipse.jetty.util.VirtualThreads.LEVEL=DEBUG
+#org.eclipse.jetty.util.thread.strategy.LEVEL=DEBUG
