+++ jetty-ee11/jetty-ee11-tests/jetty-ee11-test-integration/src/test/java/org/eclipse/jetty/ee11/test/support/StringUtil.java
@@ -13,101 +13,16 @@
 
 package org.eclipse.jetty.ee11.test.support;
 
-import java.io.BufferedReader;
-import java.io.IOException;
-import java.io.StringReader;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.ListIterator;
 
-import org.eclipse.jetty.util.IO;
+import org.eclipse.jetty.http.HttpTokens;
 
 public class StringUtil
 {
-    public static final String LN = System.getProperty("line.separator");
-
-    public static boolean isBlank(String str)
-    {
-        if (str == null)
-        {
-            return true;
-        }
-
-        int len = str.length();
-        if (len == 0)
-        {
-            return true;
-        }
-
-        char c;
-        for (int i = 0; i < str.length(); i++)
-        {
-            c = str.charAt(i);
-            if (Character.isWhitespace(c) == false)
-            {
-                return false;
-            }
-        }
-
-        return true;
-    }
-
     public static boolean isNotBlank(String str)
     {
-        return !StringUtil.isBlank(str);
-    }
-
-    public static String[] split(String s, char delim)
-    {
-        if (s == null)
-        {
-            return null;
-        }
-
-        if (s.length() <= 0)
-        {
-            return new String[]{}; // empty array
-        }
-
-        String[] ret;
-
-        int count = 0;
-        int offset = 0;
-        int idx;
-
-        // Calculate entry length to not waste memory.
-        while ((idx = s.indexOf(delim, offset)) != (-1))
-        {
-            if (idx > offset)
-            {
-                count++;
-            }
-            offset = idx + 1;
-        }
-        if (s.length() > offset)
-        {
-            count++;
-        }
-
-        // Create return array.
-        offset = 0;
-        ret = new String[count];
-        int retIdx = 0;
-        while ((idx = s.indexOf(delim, offset)) != (-1))
-        {
-            if (idx > offset)
-            {
-                ret[retIdx] = s.substring(offset, idx);
-                retIdx++;
-            }
-            offset = idx + 1;
-        }
-        if (s.length() > offset)
-        {
-            ret[retIdx] = s.substring(offset);
-        }
-
-        return ret;
+        return !org.eclipse.jetty.util.StringUtil.isBlank(str);
     }
 
     /**
@@ -133,7 +48,7 @@ public class StringUtil
                 default:
                     if (linesep)
                     {
-                        ret.append(LN);
+                        ret.append(HttpTokens.LN);
                         linesep = false;
                     }
                     ret.append(c);
@@ -141,29 +56,6 @@ public class StringUtil
         }
 
         return ret.toString();
-    }
-
-    public static List<String> asLines(String raw) throws IOException
-    {
-        List<String> lines = new ArrayList<String>();
-        StringReader sreader = null;
-        BufferedReader buf = null;
-        try
-        {
-            sreader = new StringReader(raw);
-            buf = new BufferedReader(sreader);
-            String line;
-            while ((line = buf.readLine()) != null)
-            {
-                lines.add(line);
-            }
-        }
-        finally
-        {
-            IO.close(buf);
-            IO.close(sreader);
-        }
-        return lines;
     }
 
     public static void removeStartsWith(String prefix, List<String> lines)
