+++ jetty-ee11/jetty-ee11-tests/jetty-ee11-test-integration/src/test/java/org/eclipse/jetty/ee11/test/support/rawhttp/HttpSocket.java
@@ -13,14 +13,13 @@
 
 package org.eclipse.jetty.ee11.test.support.rawhttp;
 
-import java.io.IOException;
 import java.net.InetAddress;
 import java.net.Socket;
+import java.util.function.BiFunction;
 
 /**
  * A Raw HTTP Socket connection interface.
  */
-public interface HttpSocket
+public interface HttpSocket extends BiFunction<InetAddress, Integer, Socket>
 {
-    Socket connect(InetAddress host, int port) throws IOException;
 }
