+++ jetty-ee11/jetty-ee11-tests/jetty-ee11-test-integration/src/test/java/org/eclipse/jetty/ee11/test/support/rawhttp/HttpSocketImpl.java
@@ -17,14 +17,23 @@ import java.io.IOException;
 import java.net.InetAddress;
 import java.net.Socket;
 
+import org.eclipse.jetty.io.RuntimeIOException;
+
 /**
  * Standard HTTP Socket Impl
  */
 public class HttpSocketImpl implements HttpSocket
 {
     @Override
-    public Socket connect(InetAddress host, int port) throws IOException
+    public Socket apply(InetAddress host, Integer port)
     {
-        return new Socket(host, port);
+        try
+        {
+            return new Socket(host, port);
+        }
+        catch (IOException e)
+        {
+            throw new RuntimeIOException(e);
+        }
     }
 }
