+++ jetty-ee11/jetty-ee11-tests/jetty-ee11-test-integration/src/test/java/org/eclipse/jetty/ee11/test/support/rawhttp/HttpsSocketImpl.java
@@ -22,6 +22,7 @@ import javax.net.ssl.SSLContext;
 import javax.net.ssl.SSLSocket;
 import javax.net.ssl.SSLSocketFactory;
 
+import org.eclipse.jetty.io.RuntimeIOException;
 import org.eclipse.jetty.util.ssl.SslContextFactory;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -52,11 +53,18 @@ public class HttpsSocketImpl implements HttpSocket
     }
 
     @Override
-    public Socket connect(InetAddress host, int port) throws IOException
+    public Socket apply(InetAddress host, Integer port)
     {
-        SSLSocket sslsock = (SSLSocket)sslfactory.createSocket();
-        SocketAddress address = new InetSocketAddress(host, port);
-        sslsock.connect(address);
-        return sslsock;
+        try
+        {
+            SSLSocket sslsock = (SSLSocket)sslfactory.createSocket();
+            SocketAddress address = new InetSocketAddress(host, port);
+            sslsock.connect(address);
+            return sslsock;
+        }
+        catch (IOException e)
+        {
+            throw new RuntimeIOException(e);
+        }
     }
 }
